services/client/requirements.txt
grpcio==1.60.0
protobuf